+++ src/main/java/thePackmaster/cards/evenoddpack/Flourish.java
@@ -1,12 +1,12 @@
 package thePackmaster.cards.evenoddpack;
 
-import com.megacrit.cardcrawl.cards.DamageInfo;
+import com.megacrit.cardcrawl.actions.AbstractGameAction;
+import com.megacrit.cardcrawl.actions.common.GainEnergyAction;
+import com.megacrit.cardcrawl.actions.common.HealAction;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import com.megacrit.cardcrawl.powers.PlatedArmorPower;
-import thePackmaster.actions.evenoddpack.FlourishAction;
-import thePackmaster.actions.evenoddpack.SwordAndBoardAction;
 import thePackmaster.util.Wiz;
 
 import static thePackmaster.SpireAnniversary5Mod.makeID;
@@ -64,6 +64,18 @@ public class Flourish extends AbstractEvenOddCard{
     @Override
     public void use(AbstractPlayer abstractPlayer, AbstractMonster abstractMonster) {
         Wiz.applyToSelf(new PlatedArmorPower(AbstractDungeon.player, magicNumber));
-        addToBot(new FlourishAction(magicNumber));
+        addToBot(new AbstractGameAction() {
+            @Override
+            public void update() {if (AbstractDungeon.actionManager.cardsPlayedThisTurn.size() % 2 == 0) {
+                this.addToTop(new HealAction(AbstractDungeon.player, AbstractDungeon.player, magicNumber));
+            }
+            else
+            {
+                this.addToTop(new GainEnergyAction(1));
+            }
+                this.isDone = true;
+        
+            }
+        });
     }
 }
