+++ src/main/java/thePackmaster/cards/evenoddpack/WindUpBomb.java
@@ -1,10 +1,12 @@
 package thePackmaster.cards.evenoddpack;
 
+import com.megacrit.cardcrawl.actions.AbstractGameAction;
+import com.megacrit.cardcrawl.actions.common.DamageAction;
+import com.megacrit.cardcrawl.actions.common.ModifyDamageAction;
 import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
-import thePackmaster.actions.evenoddpack.WindupAction;
 
 import static thePackmaster.SpireAnniversary5Mod.makeID;
 
@@ -72,6 +74,18 @@ public class WindUpBomb extends AbstractEvenOddCard{
     
     @Override
     public void use(AbstractPlayer abstractPlayer, AbstractMonster abstractMonster) {
-        addToBot(new WindupAction(abstractMonster, new DamageInfo(abstractMonster, this.damage, this.damageTypeForTurn), magicNumber, this));
+        addToBot(new AbstractGameAction() {
+            @Override
+            public void update() {
+                if (AbstractDungeon.actionManager.cardsPlayedThisTurn.size() % 2 == 1) {
+                    this.addToTop(new DamageAction(this.target,  new DamageInfo(abstractMonster, damage, damageTypeForTurn), AttackEffect.FIRE));
+                }
+                else
+                {
+                    this.addToTop(new ModifyDamageAction(uuid, magicNumber));
+                }
+                this.isDone = true;
+            }
+        });
     }
 }
