+++ src/main/java/thePackmaster/cards/evenoddpack/WitzBolt.java
@@ -2,15 +2,16 @@ package thePackmaster.cards.evenoddpack;
 
 import com.megacrit.cardcrawl.actions.AbstractGameAction;
 import com.megacrit.cardcrawl.actions.animations.VFXAction;
-import com.megacrit.cardcrawl.actions.common.DamageAction;
 import com.megacrit.cardcrawl.actions.utility.SFXAction;
-import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import com.megacrit.cardcrawl.powers.VulnerablePower;
+import com.megacrit.cardcrawl.powers.WeakPower;
+import com.megacrit.cardcrawl.vfx.StarBounceEffect;
 import com.megacrit.cardcrawl.vfx.combat.LightningEffect;
-import thePackmaster.actions.evenoddpack.SwordAndBoardAction;
 import thePackmaster.actions.evenoddpack.WitzBoltAction;
+import thePackmaster.util.Wiz;
 
 import static thePackmaster.SpireAnniversary5Mod.makeID;
 
@@ -66,7 +67,20 @@ public class WitzBolt extends AbstractEvenOddCard{
     @Override
     public void use(AbstractPlayer abstractPlayer, AbstractMonster abstractMonster) {
         dmg(abstractMonster, AbstractGameAction.AttackEffect.NONE);
-        addToBot(new WitzBoltAction(magicNumber, abstractMonster));
+        addToBot(new AbstractGameAction() {
+            @Override
+            public void update() {
+                if (AbstractDungeon.actionManager.cardsPlayedThisTurn.size() % 2 == 1) {
+                    Wiz.applyToEnemyTop((AbstractMonster) target, new VulnerablePower(target, magicNumber, false));
+                    Wiz.applyToEnemyTop((AbstractMonster) target, new WeakPower(target, magicNumber, false));
+                    AbstractDungeon.actionManager.addToBottom(new SFXAction("VO_CHAMP_2A", 0.5f, true));
+                    for(int i = 0; i <5; i++) {
+                        addToBot(new VFXAction(new StarBounceEffect(target.hb.cX, target.hb.cY), 0.05F));
+                    }
+                }
+                this.isDone = true;
+            }
+        });
         addToBot(new VFXAction(new LightningEffect(abstractMonster.drawX, abstractMonster.drawY), 0.05F));
     }
 }
