+++ src/main/java/thePackmaster/patches/evenoddpack/AddAuraTextPatch.java
@@ -3,17 +3,23 @@ package thePackmaster.patches.evenoddpack;
 import basemod.helpers.CardModifierManager;
 import com.evacipated.cardcrawl.modthespire.lib.*;
 import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.cards.CardGroup;
 import com.megacrit.cardcrawl.cards.DamageInfo;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import thePackmaster.cardmodifiers.evenoddpack.AuraTextModifier;
+import thePackmaster.powers.evenoddpack.GammaWardPower;
+import thePackmaster.powers.evenoddpack.PrimeDirectivePower;
 
 @SpirePatch(
-        clz = AbstractCard.class,
-        method = SpirePatch.CONSTRUCTOR,
-        paramtypez = {String.class, String.class, String.class, int.class, String.class, AbstractCard.CardType.class, AbstractCard.CardColor.class, AbstractCard.CardRarity.class, AbstractCard.CardTarget.class, DamageInfo.DamageType.class}
+        clz = CardGroup.class,
+        method = "refreshHandLayout"
 )
 public class AddAuraTextPatch {
     @SpirePostfixPatch
-    public static void AddAuraText(AbstractCard __instance, String id, String name, String imgUrl, int cost, String rawDescription, AbstractCard.CardType type, AbstractCard.CardColor color, AbstractCard.CardRarity rarity, AbstractCard.CardTarget target, DamageInfo.DamageType dType) {
-        CardModifierManager.addModifier(__instance, new AuraTextModifier());
+    public static void AddAuraText(CardGroup __instance) {
+        if(AbstractDungeon.player.hasPower(GammaWardPower.POWER_ID) || AbstractDungeon.player.hasPower(PrimeDirectivePower.POWER_ID))
+            for (AbstractCard card:AbstractDungeon.player.hand.group) {
+                CardModifierManager.addModifier(card, new AuraTextModifier());
+            }
     }
 }
